@@ -13,19 +13,18 @@
   },
   "private": true,
   "dependencies": {
-    "@agm/core": "1.0.0-beta.2",
-    "@angular/animations": "5.2.9",
-    "@angular/cdk": "5.2.4",
-    "@angular/common": "5.2.9",
-    "@angular/compiler": "5.2.9",
-    "@angular/core": "5.2.9",
-    "@angular/forms": "5.2.9",
-    "@angular/http": "5.2.9",
-    "@angular/material": "5.2.4",
-    "@angular/platform-browser": "5.2.9",
-    "@angular/platform-browser-dynamic": "5.2.9",
-    "@angular/platform-server": "5.2.9",
-    "@angular/router": "5.2.9",
+    "@angular/animations": "~7.2.0",
+    "@angular/cdk": "^7.3.3",
+    "@angular/common": "~7.2.0",
+    "@angular/compiler": "~7.2.0",
+    "@angular/core": "~7.2.0",
+    "@angular/forms": "~7.2.0",
+    "@angular/http": "~7.2.0",
+    "@angular/material": "^7.3.3",
+    "@angular/platform-browser": "~7.2.0",
+    "@angular/platform-browser-dynamic": "~7.2.0",
+    "@angular/platform-server": "~7.2.0",
+    "@angular/router": "~7.2.0",
     "@ng-bootstrap/ng-bootstrap": "2.0.0",
     "@ng-select/ng-select": "^1.4.1",
     "@types/jquery": "^3.2.12",
@@ -42,18 +41,20 @@
     "datatables.net": "^1.10.18",
     "datatables.net-dt": "^1.10.18",
     "echarts": "^4.1.0",
-    "googleapis": "28.1.0",
     "hammerjs": "2.0.8",
     "jquery": "^3.2.1",
     "jwt-decode": "^2.2.0",
     "lodash": "^4.17.10",
     "moment": "^2.19.1",
     "ng2-select": "^2.0.0",
-    "ng2-toastr": "^4.1.2",
     "ngx-easy-table": "^5.0.0-rc.1",
+    "ngx-toastr": "^9.1.2",
     "perfect-scrollbar": "1.1.0",
     "popper.js": "1.14.3",
-    "rxjs": "5.5.10",
+    "primeicons": "^1.0.0",
+    "primeng": "^7.0.5",
+    "rxjs": "~6.3.3",
+    "rxjs-compat": "~6.3.3",
     "tslib": "^1.9.0",
     "web-animations-js": "2.3.1",
     "zone.js": "0.8.4"
